Changed applying of explicit layouts to using center and bounds instead of frame.
Previously, a subview's frame was set when applying an explicit layout to an NMView and its subviews during -layoutSubviews. In case a subview has a transform which is not the identity transform, this approach can lead to unspecified behavior (the behavior of setting the frame in those cases is undefined in UIKit).

As a fix, the subviews are now repositioned and resized using the center and bounds properties whenever a layout change takes place.